debug

+++ Data Analysis/Analysis on olympic/Preprocessor.py
@@ -4,5 +4,6 @@ def preprocessor(df):
     
     df = df[df['Season']=='Summer']
     df.drop_duplicates(inplace=True)
-    df=pd.concat([df,pd.get_dummies(df['Medal']).astype('int8')],axis=1)
+    df=pd.concat([df,pd.get_dummies(df['Medal']).astype('int')],axis=1)
+    
     return df
